[optimization] Rename hir-loop-analysis into hir-loop-induction-analysis
Rename it so that the loop invariant analysis doesn't have to be in the same file.
