+++ samlang-core-optimization/__tests__/hir-loop-induction-analysis.test.ts
@@ -1,4 +1,4 @@
-import extractOptimizableWhileLoop from '../hir-loop-analysis';
+import extractOptimizableWhileLoop from '../hir-loop-induction-analysis';
 
 import {
   HIR_ZERO,
