+++ samlang-core-optimization/hir-loop-induction-analysis.ts
